+++ tasks/queue-granules/index.js
@@ -8,7 +8,6 @@ const pMap = require('p-map');
 
 const cumulusMessageAdapter = require('@cumulus/cumulus-message-adapter-js');
 const { enqueueGranuleIngestMessage } = require('@cumulus/ingest/queue');
-const { constructCollectionId } = require('@cumulus/message/Collections');
 const { buildExecutionArn } = require('@cumulus/message/Executions');
 const {
   providers: providersApi,
@@ -36,8 +35,7 @@ async function fetchGranuleProvider(prefix, providerId) {
 function groupAndBatchGranules(granules, batchSize) {
   const filteredBatchSize = isNumber(batchSize) ? batchSize : 1;
   const granulesByCollectionMap = groupBy(
-    granules,
-    (g) => constructCollectionId(g.dataType, g.version)
+    granules, granules.collectionId
   );
   const granulesBatchedByCollection = Object.values(granulesByCollectionMap).reduce(
     (arr, granulesByCollection) => arr.concat(chunk(granulesByCollection, filteredBatchSize)),
@@ -105,10 +103,7 @@ async function queueGranules(event, testMocks = {}) {
       await pMap(
         granuleBatch,
         (queuedGranule) => {
-          const collectionId = constructCollectionId(
-            queuedGranule.dataType,
-            queuedGranule.version
-          );
+          const collectionId = queuedGranule.collectionId;
 
           const granuleId = queuedGranule.granuleId;